pyproject.toml
# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["feeds_for_sendcloud/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]


# ==== black ====
[tool.black]
line-length = 119
target-version = ['py311']


# ==== isort ====
[tool.isort]
profile = "black"
line_length = 119
known_first_party = [
    "feeds_for_sendcloud",
    "config",
]
skip = ["venv/"]
skip_glob = ["**/migrations/*.py"]


# ==== mypy ====
[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"


# ==== PyLint ====
[tool.pylint.MASTER]
load-plugins = [
    "pylint_django",
    "pylint_celery",
]
django-settings-module = "config.settings.local"

[tool.pylint.FORMAT]
max-line-length = 119

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "missing-docstring",
    "invalid-name",
]

[tool.pylint.DESIGN]
max-parents = 13

[tool.pylint.TYPECHECK]
generated-members = [
    "REQUEST",
    "acl_users",
    "aq_parent",
    "[a-zA-Z]+_set{1,2}",
    "save",
    "delete",
]


# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/Riverside-Healthcare/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2


#  ==== poetry ====

[tool.poetry]
name = "feeds_for_sendcloud"
version = "0.1.0"
description = ""
authors = ["Tomás Rosales"]

[tool.poetry.dependencies]
python = "^3.9.0"
django = "==4.0.8"
django-extensions = "3.2.3"
django-environ = "==0.9.0"
django-model-utils = "==4.3.1"
djangorestframework = "==3.14.0"
django-cors-headers = "==3.13.0"
werkzeug = {version = "==2.2.2", extras = ["watchdog"]}
whitenoise = "==6.3.0"
argon2-cffi = "==21.3.0"
psycopg2-binary = "^2.9.6"
ipython = "8.12.2"
Faker = "^19.9.0"
django-filter = "^23.3"
django-redis = "5.4.0"
drf-spectacular = "0.26.5"
django-celery-beat = "2.5.0"
celery = "5.3.5"
redis = "5.0.1"
hiredis = "2.2.3"
django-lifecycle = "^1.1.2"
requests = "2.31.0"
addict = "2.4.0"
feedparser = "^6.0.10"
eventlet = "^0.33.3"

[tool.poetry.group.dev.dependencies]
ipdb = "==0.13.11"
flake8 = "==6.1.0"
flake8-isort = "==6.0.0"
black = "==22.12.0"
pylint-django = "==2.5.3"
django-debug-toolbar = "==3.8.1"
factory-boy = "3.3.0"
django-coverage-plugin = "3.1.0"
pytest-django = "4.7.0"
coverage = "7.3.2"
pylint-celery = "0.3"
pytest = "7.4.3"
pytest-sugar = "0.9.7"
mypy = "1.6.1"
djangorestframework-stubs = {version = "3.14.4", extras = ["compatible-mypy"]}
django-stubs = {version = "4.2.6", extras = ["compatible-mypy"]}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
faker = "^19.1.0"
